refactor src/components folder structure
