+++ packages/client/src/components/modals/index.js
@@ -1,0 +1,2 @@
+export * from "./GameRules";
+export * from "./GameDescription";
